requirements-streamlit.txt
# Streamlit Frontend Requirements (for Streamlit Cloud)
# Lighter version without backend dependencies

# Web Framework
streamlit>=1.28.0

# Image Processing
Pillow>=10.2.0,<11.0.0

# HTTP Requests
requests>=2.31.0

